feat: adicionando-novos-testes-e-refatorando-as-pastas-de-testes
