+++ test/widget_tests/shared/bottom_navigation_bar_test.dart
@@ -52,6 +52,7 @@ void main() {
           expect(bottomNavigationBarItemIcon, icon);
         },
       );
+
       testWidgets(
         'WHEN click BottomNavigationBarItem THEN ensure Navigates to Transactions Page',
         (tester) async {
